integrate cmn into ptarm
and separate conf.[ch] into ptarmd and ptarmcli
