+++ ptarm/misc.c
@@ -28,7 +28,7 @@
 #include <arpa/inet.h>
 
 #include "misc.h"
-#include "ptarmd.h"
+#include "ptarm_local.h"
 
 
 /**************************************************************************
@@ -105,34 +105,6 @@ void misc_datetime(char *pDateTime, size_t Len)
     time_t now = time(NULL);
     gmtime_r(&now, &tmval);
     strftime(pDateTime, Len, "%d %b %Y %T %z", &tmval);
-}
-
-
-void misc_save_event(const uint8_t *pChannelId, const char *pFormat, ...)
-{
-    char fname[256];
-
-    if (pChannelId != NULL) {
-        char chanid[LN_SZ_CHANNEL_ID * 2 + 1];
-        ptarm_util_bin2str(chanid, pChannelId, LN_SZ_CHANNEL_ID);
-        sprintf(fname, FNAME_EVENTCH_LOG, chanid);
-    } else {
-        sprintf(fname, FNAME_EVENTCH_LOG, "node");
-    }
-    FILE *fp = fopen(fname, "a");
-    if (fp != NULL) {
-        char date[50];
-        misc_datetime(date, sizeof(date));
-        fprintf(fp, "[%s]", date);
-
-        va_list ap;
-        va_start(ap, pFormat);
-        vfprintf(fp, pFormat, ap);
-        va_end(ap);
-
-        fprintf(fp, "\n");
-        fclose(fp);
-    }
 }
