+++ ptarm/misc.h
@@ -94,14 +94,6 @@ bool misc_str2bin_rev(uint8_t *pBin, uint32_t BinLen, const char *pStr);
 void misc_datetime(char *pDateTime, size_t Len);
 
 
-/** イベントファイル保存
- *
- * @param[in]       pChannelId          channel_id(ファイル名)(NULL時: "node")
- * @param[in]       pFormat             イベント文字列
- */
-void misc_save_event(const uint8_t *pChannelId, const char *pFormat, ...);
-
-
 /** 全データが0x00かのチェック
  *
  * @param[in]       pData               チェック対象
