+++ ptarmcli/Makefile
@@ -35,12 +35,10 @@ remduplicates = $(strip $(if $1,$(firstword $1) $(call remduplicates,$(filter-ou
 
 #sources project
 C_SOURCE_FILES += $(PRJ_PATH)/ptarmcli.c
-C_SOURCE_FILES += $(COMMON_PATH)/cmn/conf.c
-C_SOURCE_FILES += $(COMMON_PATH)/cmn/misc.c
+C_SOURCE_FILES += $(PRJ_PATH)/conf.c
 
 #includes common to all targets
 INC_PATHS += -I$(PRJ_PATH)
-INC_PATHS += -I$(COMMON_PATH)/cmn
 INC_PATHS += -I$(COMMON_PATH)/ptarm
 INC_PATHS += -I$(COMMON_PATH)/ptarmd
 INC_PATHS += -I$(COMMON_PATH)/libs/install/include
