+++ showdb/Makefile
@@ -3,7 +3,7 @@ OBJ = showdb
 
 include ../options.mak
 
-CFLAGS = --std=c99 -D_GNU_SOURCE -I../cmn -I../ptarm -I../ptarmd -I../libs/install/include
+CFLAGS = --std=c99 -D_GNU_SOURCE -I../ptarm -I../ptarmd -I../libs/install/include
 LDFLAGS = -L../libs/install/lib -L../ptarm
 LDFLAGS += -pthread -lptarm -llmdb -lbase58 -lmbedcrypto -lstdc++
 
@@ -18,7 +18,7 @@ endif
 all: showdb
 
 showdb: ../ptarm/libptarm.a $(SRC)
-	gcc -W -Wall $(CFLAGS) -o $(OBJ) $(SRC) ../cmn/misc.c $(LDFLAGS)
+	gcc -W -Wall $(CFLAGS) -o $(OBJ) $(SRC) $(LDFLAGS)
 
 clean:
 	-rm -rf $(OBJ)
